.tmux.conf: avoid conflicted usage of 'b' for send-keys C-b and previous-window
This is about nested tmux sessions.
If I run tmux inside tmux (and so on), I want to use C-a <n times a> to reach
n-th nested tmux session and then n/b to switch to next/previous window inside
that session.
Hence bind n/b additionally to prefix-less C-n/C-b shortcuts.

Furthermore, I want to similarly reach nested tmux sessions if they are
default-configured, so I need a key to send C-b (default) prefix as well.
I would use 'b' for similarity with 'a', but that is taken as described above.
Hence use 'm' -- it is near 'b'.