.zshrc: load early plugins before zshrc.d (thus bind fzf-tab before fzf.zsh)
This has a specific effect on my setup.

Both fzf-tab and fzf/completion.zsh (hereinafter fzf.zsh) bind their widgets
to ^I (Tab) and wrap the previously bound widget.

fzf-tab manual advises the user to load fzf-tab last (specifically, after
fzf.zsh) such that fzf-tab's widget (fzf-tab-complete) is called first and wraps
the fzf.zsh's widget (fzf-complete).

However, due to my double <Tab> hack (see `.zshrc.d/fzf`) I would prefer if the
order was reversed, such that fzf.zsh's widget is called first and wraps the
fzf-tab's widget (fzf-tab-complete).

The reason for this is that if I use a double <Tab> and fzf-complete fails to
come up with a completion, it will call the next wrapped widget, which has to be
fzf-tab-complete so that I still get some sort of fzf-powered completion instead
of expand-or-complete (the default one).