.gitconfig: set core.pager to `less -FRX` explicitly
Git normally sets $LESS to `FRX`, but only if it is not set. Since we
have $LESS set in the shell rc files, these flags are not used.
All of these flags are useful with Git, so enable them explicitly.

The downside is that we hardcode `less` here which may backfire at some
point. This is a problem for a future me though.