.config/sublime-text: Go: use #match-string-template in YAML+Go
Because we are injecting as a prototype, we will be possibly called
in the middle of other scopes, e. g. string scope.
Invoke source.go#match-string-templates to clear that scope if it exists
(thankfully, Sublime Text does not error out on scope stack underflow).

This is a hack, but oh well.