.zshrc.d: fzf: import all colors from $FZF_DEFAULT_OPTS into fzf-tab
Apparently since tmux-ftb-command spawns fzf in tmux's child process
instead of the shell's child process, it does not inherit the shell's
environment and as such does not honor $FZF_DEFAULT_OPTS (which we
export in .zshrc rather than .profile as it only matters for interactive
shells). As such, copy all colors from $FZF_DEFAULT_OPTS into fzf-flags
zstyle, not just hl= that we wanted to override.

(Actually, we should be copying the entirety of $FZF_DEFAULT_OPTS, but I
do not remember how to split variables into words in zsh so it was
easier to just hack the existing thing up...)