.zshrc.d: fzf: remove the fzf widget load order hack
We used to load fzf-tab before fzf.zsh (contrary to fzf-tab project
recommendations) to be able to do a clever hack with binding fzf.zsh
original widget to double <Tab> which then wraps fzf-tab as a fallback.

Remove the second-order hack and solve the issue properly by preventing
the infinite indirect recursion between fzf.zsh and fzf-taab, and revert
to the recommended load order for fzf-tab and fzf.zsh.