.config/bat: add a rough draft of a Solarized-ANSI color scheme
This adds a variant of a Solarized color scheme using the 16-color ANSI
palette (instead of relying on automatic "downconversion" of 24-bit RGB
colors to the generic 256-color palette, which is woefully imprecise).

As usual, this reqires that the terminal is configured with the precise
Solarized values for each of the 16 colors of the ANSI palette. The
upside is that the Solarized-ANSI color scheme is invariant with respect
to dark/light modes.