.profile: use sysfs directly instead of `nproc --all` to count CPUs
It appears that `nproc --all` reads /sys/devices/system/cpu/possible
to come up with a number, which can be inaccurate: Steam Deck has 8
logical CPUs, but /sys/devices/system/cpu/possible indicates 0-15.

Instead, read /sys/devices/system/cpu/present directly, which seems to
have the best-matching semantics. We don't parse the list fully because
there seems to be no simple way to do this, but instead just use the
last CPU index in the file (assuming that the file contents are sorted).