.config/mpv: mpv.conf: HACK: set `force-window=immediate` globally
Normally, we'd want `force-window=immediate` for network streams
(particularly for yt-dl) because there might be a huge delay between
mpv start and playback start and we want some form of user feedback
during that delay.

However, `force-window=` does not seem to work in an auto-profile, so
just set it globally. I don't use mpv for background playback anyway.